Upload

+++ src/main/java/com/emi/sprint1java/exceptions/ClientNotFoundException.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class ClientNotFoundException extends RuntimeException {
-    public ClientNotFoundException(int id) {
+    public ClientNotFoundException(long id) {
         super("Cliente não encontrado");
     }
 }
